+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Bugs
 * Fix #3408: quote pod upload file paths to support special chars
 * Fix #3561: ensure okhttp resources are closed
+* Fix #3570: added a setter for additionalProperties for proper builder support
 
 #### Improvements
 * Fix #3562: Kubernetes Mock Server improvements